requirements.txt
python-dotenv
flask
pillow
requests
types-requests
types-pillow
